requirements.txt
numpy
pandas
matplotlib.pyplot
datetime
plotly.io
plotly.express
plotly.graph_objects
plotly.figure_factory
plotly.subplots
plotly.validators.scatter.marker
matplotlib.pyplot
matplotlib.font_manager
seaborn
streamlit
FinanceDataReader
datetime
io
plotly.figures_factory